# example/pkg/api/strings/v1/strings.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "example/api/strings/v1/strings.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Strings"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/strings/info": {
      "get": {
        "operationId": "Strings_GetInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/example.api.types.v1.String"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/google.rpc.Status"
            }
          }
        },
        "parameters": [
          {
            "name": "str",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "data",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "byte"
          }
        ],
        "tags": [
          "Strings"
        ]
      }
    },
    "/api/v1/strings/upper/{str}": {
      "get": {
        "summary": "Method to upper",
        "operationId": "Strings_ToUpper",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/example.api.types.v1.String"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/google.rpc.Status"
            }
          }
        },
        "parameters": [
          {
            "name": "str",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "byte"
          }
        ],
        "tags": [
          "Strings"
        ]
      }
    }
  },
  "definitions": {
    "example.api.types.v1.String": {
      "type": "object",
      "properties": {
        "str": {
          "type": "string"
        },
        "data": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "google.protobuf.Any": {
      "type": "object",
      "properties": {
        "typeUrl": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "google.rpc.Status": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/google.protobuf.Any"
          }
        }
      }
    }
  }
}