# example/pkg/api/types/v1/types.pb.go
// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.26.0
// 	protoc        v3.17.1
// source: example/api/types/v1/types.proto

package typesV1

import (
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	structpb "google.golang.org/protobuf/types/known/structpb"
	reflect "reflect"
	sync "sync"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

type String struct {
	state         protoimpl.MessageState
	sizeCache     protoimpl.SizeCache
	unknownFields protoimpl.UnknownFields

	Str  string `protobuf:"bytes,1,opt,name=str,proto3" json:"str,omitempty"`
	Data []byte `protobuf:"bytes,2,opt,name=data,proto3" json:"data,omitempty"` // tron_option:json
}

func (x *String) Reset() {
	*x = String{}
	if protoimpl.UnsafeEnabled {
		mi := &file_example_api_types_v1_types_proto_msgTypes[0]
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		ms.StoreMessageInfo(mi)
	}
}

func (x *String) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*String) ProtoMessage() {}

func (x *String) ProtoReflect() protoreflect.Message {
	mi := &file_example_api_types_v1_types_proto_msgTypes[0]
	if protoimpl.UnsafeEnabled && x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use String.ProtoReflect.Descriptor instead.
func (*String) Descriptor() ([]byte, []int) {
	return file_example_api_types_v1_types_proto_rawDescGZIP(), []int{0}
}

func (x *String) GetStr() string {
	if x != nil {
		return x.Str
	}
	return ""
}

func (x *String) GetData() []byte {
	if x != nil {
		return x.Data
	}
	return nil
}

type StringJSON struct {
	state         protoimpl.MessageState
	sizeCache     protoimpl.SizeCache
	unknownFields protoimpl.UnknownFields

	Data *structpb.Value `protobuf:"bytes,1,opt,name=data,proto3" json:"data,omitempty"`
}

func (x *StringJSON) Reset() {
	*x = StringJSON{}
	if protoimpl.UnsafeEnabled {
		mi := &file_example_api_types_v1_types_proto_msgTypes[1]
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		ms.StoreMessageInfo(mi)
	}
}

func (x *StringJSON) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*StringJSON) ProtoMessage() {}

func (x *StringJSON) ProtoReflect() protoreflect.Message {
	mi := &file_example_api_types_v1_types_proto_msgTypes[1]
	if protoimpl.UnsafeEnabled && x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use StringJSON.ProtoReflect.Descriptor instead.
func (*StringJSON) Descriptor() ([]byte, []int) {
	return file_example_api_types_v1_types_proto_rawDescGZIP(), []int{1}
}

func (x *StringJSON) GetData() *structpb.Value {
	if x != nil {
		return x.Data
	}
	return nil
}

var File_example_api_types_v1_types_proto protoreflect.FileDescriptor

var file_example_api_types_v1_types_proto_rawDesc = []byte{
	0x0a, 0x20, 0x65, 0x78, 0x61, 0x6d, 0x70, 0x6c, 0x65, 0x2f, 0x61, 0x70, 0x69, 0x2f, 0x74, 0x79,
	0x70, 0x65, 0x73, 0x2f, 0x76, 0x31, 0x2f, 0x74, 0x79, 0x70, 0x65, 0x73, 0x2e, 0x70, 0x72, 0x6f,
	0x74, 0x6f, 0x12, 0x14, 0x65, 0x78, 0x61, 0x6d, 0x70, 0x6c, 0x65, 0x2e, 0x61, 0x70, 0x69, 0x2e,
	0x74, 0x79, 0x70, 0x65, 0x73, 0x2e, 0x76, 0x31, 0x1a, 0x1c, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65,
	0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2f, 0x73, 0x74, 0x72, 0x75, 0x63, 0x74,
	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0x2e, 0x0a, 0x06, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67,
	0x12, 0x10, 0x0a, 0x03, 0x73, 0x74, 0x72, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x73,
	0x74, 0x72, 0x12, 0x12, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x61, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0c,
	0x52, 0x04, 0x64, 0x61, 0x74, 0x61, 0x22, 0x38, 0x0a, 0x0a, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67,
	0x4a, 0x53, 0x4f, 0x4e, 0x12, 0x2a, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x61, 0x18, 0x01, 0x20, 0x01,
	0x28, 0x0b, 0x32, 0x16, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74,
	0x6f, 0x62, 0x75, 0x66, 0x2e, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x04, 0x64, 0x61, 0x74, 0x61,
	0x42, 0x22, 0x5a, 0x20, 0x65, 0x78, 0x61, 0x6d, 0x70, 0x6c, 0x65, 0x2f, 0x70, 0x6b, 0x67, 0x2f,
	0x61, 0x70, 0x69, 0x2f, 0x74, 0x79, 0x70, 0x65, 0x73, 0x2f, 0x76, 0x31, 0x3b, 0x74, 0x79, 0x70,
	0x65, 0x73, 0x56, 0x31, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
}

var (
	file_example_api_types_v1_types_proto_rawDescOnce sync.Once
	file_example_api_types_v1_types_proto_rawDescData = file_example_api_types_v1_types_proto_rawDesc
)

func file_example_api_types_v1_types_proto_rawDescGZIP() []byte {
	file_example_api_types_v1_types_proto_rawDescOnce.Do(func() {
		file_example_api_types_v1_types_proto_rawDescData = protoimpl.X.CompressGZIP(file_example_api_types_v1_types_proto_rawDescData)
	})
	return file_example_api_types_v1_types_proto_rawDescData
}

var file_example_api_types_v1_types_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
var file_example_api_types_v1_types_proto_goTypes = []interface{}{
	(*String)(nil),         // 0: example.api.types.v1.String
	(*StringJSON)(nil),     // 1: example.api.types.v1.StringJSON
	(*structpb.Value)(nil), // 2: google.protobuf.Value
}
var file_example_api_types_v1_types_proto_depIdxs = []int32{
	2, // 0: example.api.types.v1.StringJSON.data:type_name -> google.protobuf.Value
	1, // [1:1] is the sub-list for method output_type
	1, // [1:1] is the sub-list for method input_type
	1, // [1:1] is the sub-list for extension type_name
	1, // [1:1] is the sub-list for extension extendee
	0, // [0:1] is the sub-list for field type_name
}

func init() { file_example_api_types_v1_types_proto_init() }
func file_example_api_types_v1_types_proto_init() {
	if File_example_api_types_v1_types_proto != nil {
		return
	}
	if !protoimpl.UnsafeEnabled {
		file_example_api_types_v1_types_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
			switch v := v.(*String); i {
			case 0:
				return &v.state
			case 1:
				return &v.sizeCache
			case 2:
				return &v.unknownFields
			default:
				return nil
			}
		}
		file_example_api_types_v1_types_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
			switch v := v.(*StringJSON); i {
			case 0:
				return &v.state
			case 1:
				return &v.sizeCache
			case 2:
				return &v.unknownFields
			default:
				return nil
			}
		}
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: file_example_api_types_v1_types_proto_rawDesc,
			NumEnums:      0,
			NumMessages:   2,
			NumExtensions: 0,
			NumServices:   0,
		},
		GoTypes:           file_example_api_types_v1_types_proto_goTypes,
		DependencyIndexes: file_example_api_types_v1_types_proto_depIdxs,
		MessageInfos:      file_example_api_types_v1_types_proto_msgTypes,
	}.Build()
	File_example_api_types_v1_types_proto = out.File
	file_example_api_types_v1_types_proto_rawDesc = nil
	file_example_api_types_v1_types_proto_goTypes = nil
	file_example_api_types_v1_types_proto_depIdxs = nil
}